Seccion 6 - 48 - Implementacion modulo components - Implementacion componente incrementador
