+++ adminpro/src/app/components/incrementador/incrementador.component.html
@@ -1,0 +1,13 @@
+<div class="input-group">
+    <span class="input-group-btn">
+        <button class="btn btn-primary" type="button" (click)="cambiarValor(-5)">
+            <i class="fa fa-minus"></i>
+        </button>
+    </span>
+    <input type="number" class="form-control" placeholder="Progreso" [(ngModel)]="progreso">
+    <span class="input-group-btn">
+        <button class="btn btn-primary" type="button" (click)="cambiarValor(5)">
+            <i class="fa fa-plus"></i>
+        </button>
+    </span>
+</div>
