+++ adminpro/src/app/pages/pages.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
-import { FormsModule } from "@angular/forms";
 
 import { SharedModule } from '../shared/shared.module';
 
@@ -9,6 +8,7 @@ import { DashboardComponent } from './dashboard/dashboard.component';
 import { ProgressComponent } from './progress/progress.component';
 import { Grafica1Component } from './grafica1/grafica1.component';
 import { PagesComponent } from './pages.component';
+import { ComponentsModule } from '../components/components.module';
 
 
 @NgModule({
@@ -27,8 +27,8 @@ import { PagesComponent } from './pages.component';
   imports: [
     CommonModule,
     RouterModule,
-    FormsModule,
-    SharedModule
+    SharedModule,
+    ComponentsModule
   ]
 })
 export class PagesModule { }
