+++ adminpro/src/app/pages/progress/progress.component.html
@@ -6,8 +6,15 @@
                 <div class="progress m-t-20">
                     <div class="progress-bar bg-primary progress-bar-striped active" 
                          style="height:10px;"
-                         [style.width]="getPorcentaje"
                          role="progressbar">
+                         <!-- [style.width]="getPorcentaje" -->
+                    </div>
+                </div>
+                <div class="progress m-t-20">
+                    <div class="progress-bar bg-primary progress-bar-striped active" 
+                         style="height:10px;"
+                         role="progressbar">
+                         <!-- [style.width]="getPorcentaje" -->
                     </div>
                 </div>
             </div>
@@ -19,26 +26,14 @@
     <div class="col-6">
         <div class="card">
             <div class="card-body">
-                <div class="input-group">
-                    <span class="input-group-btn">
-                        <button class="btn btn-primary" type="button" (click)="cambiarValor(-5)">
-                            <i class="fa fa-minus"></i>
-                        </button>
-                    </span>
-                    <input type="number" class="form-control" placeholder="Progreso" [(ngModel)]="progreso">
-                    <span class="input-group-btn">
-                        <button class="btn btn-primary" type="button" (click)="cambiarValor(5)">
-                            <i class="fa fa-plus"></i>
-                        </button>
-                    </span>
-                </div>
+                <app-incrementador></app-incrementador>
             </div>
         </div>
     </div>
     <div class="col-6">
         <div class="card">
             <div class="card-body">
-                2
+                <app-incrementador></app-incrementador>
             </div>
         </div>
     </div>
