+++ adminpro/src/app/pages/progress/progress.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-progress',
@@ -7,16 +7,6 @@ import { Component, OnInit } from '@angular/core';
 })
 export class ProgressComponent {
   
-  progreso: number = 10;
 
-  get getPorcentaje() {
-    return `${ this.progreso }%`
-  }
-
-  cambiarValor( valor: number ){
-    this.progreso = this.progreso + valor;
-    if(this.progreso>=100 && valor>=0) this.progreso = 100;
-    if(this.progreso<=0 && valor<0) this.progreso = 0;
-  }
 
 }
